+++ src/app/review/_component/SelectReview.tsx
@@ -1,5 +1,5 @@
 import { useEffect, useState } from "react";
-import { Chip, Chips } from "../_component/common/Chips";
+import { Chip, Chips } from "../../_component/common/Chips";
 
 interface SelectedReviewProps {
   onSelected: (value: string[]) => void;
